Fixed build. Broken Quiz & Contact Edit fixed.
The support network edit was broken so I commented out the code.
The QuizServiceTest was relying on some functionality not present in QuizService.
Added the method (with a runtime exception to be implemented later).
Also the ScoresDBOpenHelper had been renamed so I fixed the file to be the same name.

Also committing my dialog code as it got caught into the commit when I pulled.  It
might be valuable for John to use.