Database crashes fixed. Refactored db helpers.
Apparently only one onCreate method for a dbHelper (with the same database)
can exist.  Either you have separate databases or you need to have a single dbHelper.

Moved the logic of ScoresDBOpenHelper and SupportContactOpenHelper to PCSDBOpenHelper.
Then the scores and support contact helper become table definition classes.

Since the table definitions are in a number of classes this had an impact on a lot of classes.