Highlighting bug fix w/ deletion.
If a crisis contact is setup and highlighted inbetween two other support contacts
and it was deleted.  The same view was being re-used and a non-crisis contact
would then be highlighted as the crisis one.

We had to setup the views to be re-created when alternating on the adapter.
This fixes the bug.