Library Resource Controller UI test stubs.
Made it so you can navigate to the 4 different types of library resource activities from
the library controller.

Implemented the LibraryResourceListAdapter so we can display the list of library resources.

This accomplishes the prep work for the library to load it's resources from a JSON file and
have them work automatically with the library controller.