Library controller gets resources from service.
Hooked up the library controller so that it uses the service observer class and retrieves it's resources
from the resource library service.  As we don't have the JSON rest service built yet I implemented a fake
http json loader that will just return dummy json values that the service relies on.

Everything is wired up together now to fill in the deserializer details (title, description, etc)
and to hook up to the actual JSON api using the HTTPJsonLoader