Added IntentChecker to MeaningfulPicture intent.
So we don't get any application crashes I added the IntentChecker
to the meaningful picture stetings controller.  Since this requires
popping up a dialog to let the user know they need to install the application,
I refactored the dialogerHelper from the support network list controller
into a DialogHelper class in utils that then could be used by the settings
controller.