Implemented FirstUseChecklistService.
We were faking the setup marking the use complete even though a user hadn't actually finished the setup.

I copied the implementation from FirstUseController into the service and set up the various setup controllers
to mark the service as being complete and return to the FirstUseController.  This makes it so the items
actually must be completed before the setup process can be finished.